fix(web): expand hidden projects below their disclosure
